refactor: inline CommandLineValidator into Cli.Pipeline
Closes #219
