+++ src/Program.cs
@@ -21,7 +21,8 @@ namespace Zipper
                 }
                 catch (Exception ex)
                 {
-                    Console.Error.WriteLine($"\nBenchmark error: {ex.Message}");
+                    Console.Error.WriteLine($"
+Benchmark error: {ex.Message}");
                     return 1;
                 }
             }
@@ -32,11 +33,10 @@ namespace Zipper
                 return 0;
             }
 
-            // Validate and parse command line arguments
-            var request = CommandLineValidator.ValidateAndParseArguments(args);
+            var request = Cli.Pipeline.Build(args);
             if (request == null)
             {
-                return 1; // Error already displayed by CommandLineValidator
+                return 1;
             }
 
             IGenerationMode mode = SelectMode(request);
